Reemit 'end' event from socket when using MCCP both locally and remote
The event 'end' is never emitted when listening from a client, using MCCP on remote and local.
This is a bit hacky, but I didn't find any other solution.
Should have been done in MCCP.js, but didn't find any good way.
I don't really know what the real problem is, could be the stream is in the
middle of expecting to decompress when it receives an 'end', and it will be ignored?